pyproject.toml
[project]
name = "netunicorn-library"
version = "0.3.2"
authors = [
    {name = "Roman Beltiukov", email = "rbeltiukov@ucsb.edu"},
]
description = "netunicorn contrib library"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["netunicorn"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "netunicorn-base >= 0.3.1",
]

[tool.setuptools.packages.find]
where = ["src"]